+++ src/main/scala/edu/ie3/simona/model/participant/load/LoadReference.scala
@@ -7,7 +7,7 @@
 package edu.ie3.simona.model.participant.load
 
 import edu.ie3.datamodel.models.input.system.LoadInput
-import edu.ie3.simona.config.SimonaConfig
+import edu.ie3.simona.config.RuntimeConfig.LoadRuntimeConfig
 import edu.ie3.util.StringUtils
 import edu.ie3.util.quantities.PowerSystemUnits.{MEGAWATT, MEGAWATTHOUR}
 import squants.energy.{MegawattHours, Megawatts}
@@ -69,7 +69,7 @@ object LoadReference {
     */
   def apply(
       inputModel: LoadInput,
-      modelConfig: SimonaConfig.LoadRuntimeConfig,
+      modelConfig: LoadRuntimeConfig,
   ): LoadReference =
     StringUtils.cleanString(modelConfig.reference).toLowerCase match {
       case "power" =>