+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentIT.scala
@@ -13,7 +13,12 @@ import edu.ie3.simona.agent.participant.load.LoadAgent.FixedLoadAgent
 import edu.ie3.simona.agent.participant.pv.PvAgent
 import edu.ie3.simona.agent.participant.statedata.ParticipantStateData.ParticipantInitializeStateData
 import edu.ie3.simona.agent.participant.storage.StorageAgent
-import edu.ie3.simona.config.SimonaConfig._
+import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+  DataProvision,
+  RegistrationFailedMessage,
+  RegistrationSuccessfulMessage,
+}
+import edu.ie3.simona.config.RuntimeConfig._
 import edu.ie3.simona.event.ResultEvent
 import edu.ie3.simona.event.ResultEvent.ParticipantResultEvent
 import edu.ie3.simona.event.notifier.NotifierConfig
@@ -23,12 +28,7 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.{
 }
 import edu.ie3.simona.ontology.messages.services.ServiceMessage
 import edu.ie3.simona.ontology.messages.services.ServiceMessage.PrimaryServiceRegistrationMessage
-import edu.ie3.simona.ontology.messages.services.ServiceMessage.RegistrationResponseMessage.{
-  RegistrationFailedMessage,
-  RegistrationSuccessfulMessage,
-}
 import edu.ie3.simona.ontology.messages.services.WeatherMessage.{
-  ProvideWeatherMessage,
   RegisterForWeatherMessage,
   WeatherData,
 }
@@ -62,7 +62,6 @@ class EmAgentIT
     with should.Matchers
     with EmInputTestData
     with MockitoSugar {
-
   // start a bit later so the sun is up
   protected implicit val simulationStartDate: ZonedDateTime =
     TimeUtil.withDefaults.toZonedDateTime("2020-01-01T10:00:00Z")
@@ -85,11 +84,8 @@ class EmAgentIT
   )
 
   override protected val modelConfig: EmRuntimeConfig = EmRuntimeConfig(
-    calculateMissingReactivePowerWithModel = false,
-    scaling = 1d,
     uuids = List("default"),
     aggregateFlex = "SELF_OPT",
-    curtailRegenerative = false,
   )
 
   private implicit val quantityTolerance: Double = 1e-10d
@@ -124,11 +120,7 @@ class EmAgentIT
             initStateData = ParticipantInitializeStateData(
               loadInput,
               LoadRuntimeConfig(
-                calculateMissingReactivePowerWithModel = true,
-                scaling = 1d,
-                modelBehaviour = "fix",
-                reference = "power",
-                uuids = List.empty,
+                calculateMissingReactivePowerWithModel = true
               ),
               primaryServiceProxy.ref.toClassic,
               None,
@@ -150,7 +142,6 @@ class EmAgentIT
               pvInput,
               PvRuntimeConfig(
                 calculateMissingReactivePowerWithModel = true,
-                scaling = 2d,
                 uuids = List.empty,
               ),
               primaryServiceProxy.ref.toClassic,
@@ -173,9 +164,7 @@ class EmAgentIT
               householdStorageInput,
               StorageRuntimeConfig(
                 calculateMissingReactivePowerWithModel = true,
-                scaling = 1d,
                 uuids = List.empty,
-                initialSoc = 0d,
                 targetSoc = None,
               ),
               primaryServiceProxy.ref.toClassic,
@@ -200,7 +189,7 @@ class EmAgentIT
         loadAgent ! Activation(INIT_SIM_TICK)
 
         primaryServiceProxy.expectMessage(
-          PrimaryServiceRegistrationMessage(loadInput.getUuid)
+          PrimaryServiceRegistrationMessage(loadAgent.ref, loadInput.getUuid)
         )
         loadAgent ! RegistrationFailedMessage(primaryServiceProxy.ref.toClassic)
 
@@ -227,7 +216,7 @@ class EmAgentIT
         pvAgent ! Activation(INIT_SIM_TICK)
 
         primaryServiceProxy.expectMessage(
-          PrimaryServiceRegistrationMessage(pvInput.getUuid)
+          PrimaryServiceRegistrationMessage(pvAgent.ref, pvInput.getUuid)
         )
         pvAgent ! RegistrationFailedMessage(primaryServiceProxy.ref.toClassic)
 
@@ -241,7 +230,7 @@ class EmAgentIT
 
         pvAgent ! RegistrationSuccessfulMessage(
           weatherService.ref.toClassic,
-          Some(0L),
+          0L,
         )
 
         scheduler.expectMessage(Completion(pvAgent))
@@ -250,7 +239,10 @@ class EmAgentIT
         storageAgent ! Activation(INIT_SIM_TICK)
 
         primaryServiceProxy.expectMessage(
-          PrimaryServiceRegistrationMessage(householdStorageInput.getUuid)
+          PrimaryServiceRegistrationMessage(
+            storageAgent.ref,
+            householdStorageInput.getUuid,
+          )
         )
         storageAgent ! RegistrationFailedMessage(
           primaryServiceProxy.ref.toClassic
@@ -260,20 +252,20 @@ class EmAgentIT
 
         /* TICK 0
          LOAD: 0.269 kW
-         PV:  -5.617 kW
+         PV:  -5.842 kW
          STORAGE: SOC 0 %
          -> charge with 5 kW
-         -> remaining -0.348 kW
+         -> remaining -0.573 kW
          */
 
         emAgentActivation ! Activation(0)
 
-        pvAgent ! ProvideWeatherMessage(
+        pvAgent ! DataProvision(
           0,
           weatherService.ref.toClassic,
           WeatherData(
-            WattsPerSquareMeter(540d),
             WattsPerSquareMeter(200d),
+            WattsPerSquareMeter(100d),
             Celsius(0d),
             MetersPerSecond(0d),
           ),
@@ -285,29 +277,31 @@ class EmAgentIT
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 0L.toDateTime
             emResult.getP should equalWithTolerance(
-              -0.00034885012.asMegaWatt
+              -0.00057340027059.asMegaWatt
             )
-            emResult.getQ should equalWithTolerance(0.00008828554.asMegaVar)
+            emResult.getQ should equalWithTolerance(
+              0.0000882855367033.asMegaVar
+            )
         }
 
         scheduler.expectMessage(Completion(emAgentActivation, Some(7200)))
 
         /* TICK 7200
          LOAD: 0.269 kW (unchanged)
-         PV:  -3.651 kW
+         PV:  -3.791 kW
          STORAGE: SOC 63.3 %
-         -> charge with 3.382 kW
+         -> charge with 3.522 kW
          -> remaining 0 kW
          */
 
         emAgentActivation ! Activation(7200)
 
-        pvAgent ! ProvideWeatherMessage(
+        pvAgent ! DataProvision(
           7200,
           weatherService.ref.toClassic,
           WeatherData(
-            WattsPerSquareMeter(300d),
-            WattsPerSquareMeter(500d),
+            WattsPerSquareMeter(50d),
+            WattsPerSquareMeter(150d),
             Celsius(0d),
             MetersPerSecond(0d),
           ),
@@ -318,28 +312,30 @@ class EmAgentIT
           case ParticipantResultEvent(emResult: EmResult) =>
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 7200.toDateTime
-            emResult.getP should equalWithTolerance(0.asMegaWatt)
+            emResult.getP should equalWithTolerance(
+              0.0.asMegaWatt
+            )
             emResult.getQ should equalWithTolerance(0.0000882855367.asMegaVar)
         }
 
-        scheduler.expectMessage(Completion(emAgentActivation, Some(13362)))
+        scheduler.expectMessage(Completion(emAgentActivation, Some(13115)))
 
-        /* TICK 13362
+        /* TICK 13115
          LOAD: 0.269 kW (unchanged)
-         PV:  -3.651 kW (unchanged)
+         PV:  -3.791 kW (unchanged)
          STORAGE: SOC 100 %
          -> charge with 0 kW
-         -> remaining -3.382 kW
+         -> remaining -3.522 kW
          */
 
-        emAgentActivation ! Activation(13362)
+        emAgentActivation ! Activation(13115)
 
         resultListener.expectMessageType[ParticipantResultEvent] match {
           case ParticipantResultEvent(emResult: EmResult) =>
             emResult.getInputModel shouldBe emInput.getUuid
-            emResult.getTime shouldBe 13362L.toDateTime
+            emResult.getTime shouldBe 13115L.toDateTime
             emResult.getP should equalWithTolerance(
-              -0.003382375474.asMegaWatt
+              -0.0035233186089842434.asMegaWatt
             )
             emResult.getQ should equalWithTolerance(0.0000882855367.asMegaVar)
         }
@@ -348,20 +344,20 @@ class EmAgentIT
 
         /* TICK 14400
          LOAD: 0.269 kW (unchanged)
-         PV:  -0.066 kW
+         PV:  -0.069 kW
          STORAGE: SOC 100 %
-         -> charge with -0.202956 kW
-         -> remaining 0 kW
+         -> discharge with 0.2 kW
+         -> remaining 0.0 kW
          */
 
         // send weather data before activation, which can happen
         // it got cloudy now...
-        pvAgent ! ProvideWeatherMessage(
+        pvAgent ! DataProvision(
           14400,
           weatherService.ref.toClassic,
           WeatherData(
-            WattsPerSquareMeter(5d),
-            WattsPerSquareMeter(5d),
+            WattsPerSquareMeter(0.5d),
+            WattsPerSquareMeter(2d),
             Celsius(0d),
             MetersPerSecond(0d),
           ),
@@ -374,8 +370,10 @@ class EmAgentIT
           case ParticipantResultEvent(emResult: EmResult) =>
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 14400L.toDateTime
-            emResult.getP should equalWithTolerance(0.asMegaWatt)
-            emResult.getQ should equalWithTolerance(0.000088285537.asMegaVar)
+            emResult.getP should equalWithTolerance(
+              0.0.asMegaWatt
+            )
+            emResult.getQ should equalWithTolerance(0.000088285536.asMegaVar)
         }
 
         scheduler.expectMessage(Completion(emAgentActivation, Some(21600)))
@@ -410,11 +408,7 @@ class EmAgentIT
             initStateData = ParticipantInitializeStateData(
               loadInput,
               LoadRuntimeConfig(
-                calculateMissingReactivePowerWithModel = true,
-                scaling = 1d,
-                modelBehaviour = "fix",
-                reference = "power",
-                uuids = List.empty,
+                calculateMissingReactivePowerWithModel = true
               ),
               primaryServiceProxy.ref.toClassic,
               None,
@@ -436,7 +430,6 @@ class EmAgentIT
               pvInput,
               PvRuntimeConfig(
                 calculateMissingReactivePowerWithModel = true,
-                scaling = 2d,
                 uuids = List.empty,
               ),
               primaryServiceProxy.ref.toClassic,
@@ -485,7 +478,7 @@ class EmAgentIT
         loadAgent ! Activation(INIT_SIM_TICK)
 
         primaryServiceProxy.expectMessage(
-          PrimaryServiceRegistrationMessage(loadInput.getUuid)
+          PrimaryServiceRegistrationMessage(loadAgent.ref, loadInput.getUuid)
         )
         loadAgent ! RegistrationFailedMessage(primaryServiceProxy.ref.toClassic)
 
@@ -512,7 +505,7 @@ class EmAgentIT
         pvAgent ! Activation(INIT_SIM_TICK)
 
         primaryServiceProxy.expectMessage(
-          PrimaryServiceRegistrationMessage(pvInput.getUuid)
+          PrimaryServiceRegistrationMessage(pvAgent.ref, pvInput.getUuid)
         )
         pvAgent ! RegistrationFailedMessage(primaryServiceProxy.ref.toClassic)
 
@@ -526,7 +519,7 @@ class EmAgentIT
 
         pvAgent ! RegistrationSuccessfulMessage(
           weatherService.ref.toClassic,
-          Some(0L),
+          0L,
         )
 
         scheduler.expectMessage(Completion(pvAgent))
@@ -535,7 +528,10 @@ class EmAgentIT
         heatPumpAgent ! Activation(INIT_SIM_TICK)
 
         primaryServiceProxy.expectMessage(
-          PrimaryServiceRegistrationMessage(adaptedHpInputModel.getUuid)
+          PrimaryServiceRegistrationMessage(
+            heatPumpAgent.ref,
+            adaptedHpInputModel.getUuid,
+          )
         )
         heatPumpAgent ! RegistrationFailedMessage(
           primaryServiceProxy.ref.toClassic
@@ -543,14 +539,14 @@ class EmAgentIT
 
         weatherService.expectMessage(
           RegisterForWeatherMessage(
-            hpInputModel.getNode.getGeoPosition.getY,
-            hpInputModel.getNode.getGeoPosition.getX,
+            adaptedHpInputModel.getNode.getGeoPosition.getY,
+            adaptedHpInputModel.getNode.getGeoPosition.getX,
           )
         )
 
         heatPumpAgent ! RegistrationSuccessfulMessage(
           weatherService.ref.toClassic,
-          Some(0L),
+          0L,
         )
 
         scheduler.expectMessage(Completion(heatPumpAgent))
@@ -559,21 +555,21 @@ class EmAgentIT
 
         /* TICK 0
          LOAD: 0.269 kW
-         PV:  -5.617 kW
+         PV:  -5.842 kW
          Heat pump: off, can be turned on or stay off
          -> set point ~3.5 kW (bigger than 50 % rated apparent power): turned on
-         -> remaining -0.499 kW
+         -> remaining -0.723 kW
          */
 
         emAgentActivation ! Activation(0)
 
         weatherDependentAgents.foreach {
-          _ ! ProvideWeatherMessage(
+          _ ! DataProvision(
             0,
             weatherService.ref.toClassic,
             WeatherData(
-              WattsPerSquareMeter(540d),
               WattsPerSquareMeter(200d),
+              WattsPerSquareMeter(100d),
               Celsius(0d),
               MetersPerSecond(0d),
             ),
@@ -586,30 +582,32 @@ class EmAgentIT
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 0.toDateTime
             emResult.getP should equalWithTolerance(
-              -0.000498850118.asMegaWatt
+              -0.0007234002705905523.asMegaWatt
             )
-            emResult.getQ should equalWithTolerance(0.001073120041.asMegaVar)
+            emResult.getQ should equalWithTolerance(
+              0.0010731200407782782.asMegaVar
+            )
         }
 
         scheduler.expectMessage(Completion(emAgentActivation, Some(7200)))
 
         /* TICK 7200
          LOAD: 0.269 kW (unchanged)
-         PV:  -3.651 kW
+         PV:  -3.791 kW
          Heat pump: running (turned on from last request), can also be turned off
          -> set point ~3.5 kW (bigger than 50 % rated apparent power): stays turned on with unchanged state
-         -> remaining 1.468 kW
+         -> remaining 1.327 kW
          */
 
         emAgentActivation ! Activation(7200)
 
         weatherDependentAgents.foreach {
-          _ ! ProvideWeatherMessage(
+          _ ! DataProvision(
             7200,
             weatherService.ref.toClassic,
             WeatherData(
-              WattsPerSquareMeter(300d),
-              WattsPerSquareMeter(500d),
+              WattsPerSquareMeter(50d),
+              WattsPerSquareMeter(150d),
               Celsius(0d),
               MetersPerSecond(0d),
             ),
@@ -621,30 +619,34 @@ class EmAgentIT
           case ParticipantResultEvent(emResult: EmResult) =>
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 7200.toDateTime
-            emResult.getP should equalWithTolerance(0.001467624526.asMegaWatt)
-            emResult.getQ should equalWithTolerance(0.001073120041.asMegaVar)
+            emResult.getP should equalWithTolerance(
+              0.0013266813910157566.asMegaWatt
+            )
+            emResult.getQ should equalWithTolerance(
+              0.0010731200407782782.asMegaVar
+            )
         }
 
         scheduler.expectMessage(Completion(emAgentActivation, Some(14400)))
 
         /* TICK 14400
          LOAD: 0.269 kW (unchanged)
-         PV:  -0.066 kW
+         PV:  -0.07 kW
          Heat pump: Is still running, can still be turned off
          -> flex signal is 0 MW: Heat pump is turned off
-         -> remaining 0.203 kW
+         -> remaining 0.199 kW
          */
 
         emAgentActivation ! Activation(14400)
 
         // it got cloudy now...
         weatherDependentAgents.foreach {
-          _ ! ProvideWeatherMessage(
+          _ ! DataProvision(
             14400,
             weatherService.ref.toClassic,
             WeatherData(
-              WattsPerSquareMeter(5d),
-              WattsPerSquareMeter(5d),
+              WattsPerSquareMeter(0.5d),
+              WattsPerSquareMeter(2d),
               Celsius(0d),
               MetersPerSecond(0d),
             ),
@@ -656,30 +658,34 @@ class EmAgentIT
           case ParticipantResultEvent(emResult: EmResult) =>
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 14400L.toDateTime
-            emResult.getP should equalWithTolerance(0.000202956264.asMegaWatt)
-            emResult.getQ should equalWithTolerance(0.000088285537.asMegaVar)
+            emResult.getP should equalWithTolerance(
+              0.00019892577822992104.asMegaWatt
+            )
+            emResult.getQ should equalWithTolerance(
+              0.0000882855367033582.asMegaVar
+            )
         }
 
         scheduler.expectMessage(Completion(emAgentActivation, Some(21600)))
 
         /* TICK 21600
          LOAD: 0.269 kW (unchanged)
-         PV:  -0.026 kW
+         PV:  -0.023 kW
          Heat pump: Is not running, can run or stay off
          -> flex signal is 0 MW: Heat pump is turned off
-         -> remaining 0.242 kW
+         -> remaining 0.245 kW
          */
 
         emAgentActivation ! Activation(21600)
 
         weatherDependentAgents.foreach {
-          _ ! ProvideWeatherMessage(
+          _ ! DataProvision(
             21600,
             weatherService.ref.toClassic,
             WeatherData(
               // Same irradiation, but different angle of the sun
-              WattsPerSquareMeter(5d),
-              WattsPerSquareMeter(5d),
+              WattsPerSquareMeter(2d),
+              WattsPerSquareMeter(4d),
               Celsius(0d),
               MetersPerSecond(0d),
             ),
@@ -691,8 +697,12 @@ class EmAgentIT
           case ParticipantResultEvent(emResult: EmResult) =>
             emResult.getInputModel shouldBe emInput.getUuid
             emResult.getTime shouldBe 21600.toDateTime
-            emResult.getP should equalWithTolerance(0.000242284024.asMegaWatt)
-            emResult.getQ should equalWithTolerance(0.000088285537.asMegaVar)
+            emResult.getP should equalWithTolerance(
+              0.0002450436827011999.asMegaWatt
+            )
+            emResult.getQ should equalWithTolerance(
+              0.0000882855367033582.asMegaVar
+            )
         }
 
         scheduler.expectMessage(Completion(emAgentActivation, Some(28800)))